split reviewing page into different components
