+++ src/lib/components/reviewing/ReviewDecision.tsx
@@ -1,0 +1,12 @@
+'use client';
+
+import type { TaskWithReviewData } from '@/lib/types/position.types';
+
+type ReviewDecisionProps = {
+    task: TaskWithReviewData | null;
+};
+
+// TODO: accept/reject decision panel
+export default function ReviewDecision(_props: ReviewDecisionProps) {
+    return <div>TODO</div>;
+}
